override mysqljs type casting for DateTime
This fixes a problem where timezones are returned from the API
with "Z" at the end of them, causing them to always be parsed
by end-user libraries as UTC time.

Currently this is not a problem if the end user is also in the UTC
timezone, but otherwise it will cause the time to drift by the
offset from UTC the user is currently in.